+++ env/src/lib.rs
@@ -1,8 +1,11 @@
 mod config;
+mod diagnostic;
 mod dir;
 mod file;
 mod file_system;
 mod media;
 mod similarity;
 
-pub use {config::settings::*, dir::*, file::*, file_system::*, media::*, similarity::*};
+pub use {
+    config::settings::*, diagnostic::*, dir::*, file::*, file_system::*, media::*, similarity::*,
+};